requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
requests>=2.31.0
databricks-sdk>=0.20.0
apscheduler>=3.10.4
